tool/requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
mutagen>=1.47.0
pymediainfo>=6.0.0
